Do not overflow local buffer
Do not use full width of buf, as we intend to assign

    buf[num_read] = '\0';

at the end of the buffer, afterwards. In case we have just read full BUF_SIZE,
we will write '\0' to a memory location that is outside of buf. This can be
referred to as CWE-119 although, it seems pretty harmless in this case.

Signed-off-by: Šimon Lukašík <isimluk@fedoraproject.org>