Fix Makefile to rebuild when sources are touched
Dependencies were not fully specified, so touching a source file didn't
trigger a rebuild and `make clean` was necessary; fix that.

Furthermore, use $(CFLAGS) and $(LDFLAGS) correctly. Don't pass a header
file to the linker step, use a dependency-order rule (GNU make specific)
for creating the target bin/ directory and mark a few more rules as
.PHONY:.

Tested that the produced binary is identical, touching source files and
headers properly rebuilds object files and the final binary and that
`make clean` removes the temporary files too.

Signed-off-by: Filipe Brandenburger <filbranden@google.com>