Add 'trace' log level support
The trace log level is currently not supported by conmon, but seems to be
a valid level for container runtimes like CRI-O and podman. This results
that the container currently won't start if the log level is set to
'trace'. Beside this, we now case insensitively compare the log level
string to make it more fail safe.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>