conmon: kill the process group on timeout
when the timeout is expired, instead of killing the process, make sure
to kill the entire process group.

It is not safe as using pid cgroup or a pid namespace, which we cannot
use as the script must run inside of the container pid namespace, but
it addresses most common cases like bash forking a new process.

Closes: https://bugzilla.redhat.com/show_bug.cgi?id=1774184

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>