conmon: check oom counter on cgroup v1
when running on cgroup v1, also check whether the OOM counter is > 0
before writing the "oom" file.

Previously, conmon was detecting an OOM every time an event fd was
triggered.

Unfortunately that is not enough, as an event on the eventfd is also
triggered when the cgroup is deleted.  That could cause a race
condition where conmon sees the event from the cgroup deletion before
the SIGCHLD+SIGUSR1.  That could happen for example when systemd
CollectMode is set for failed units.

Closes: https://bugzilla.redhat.com/show_bug.cgi?id=1782601

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>