oom v1: verify cgroup hasn't been cleaned up
This PR is a replacement for https://github.com/containers/conmon/commit/4100fb20bc156fdd602f8684affa27e6b139d0ad, in which we verify an OOM occured when we get triggered on the eventfd.

Instead of the aforementioned fix, which read the memory.oom_control file, and verified it correctly contained oom_kill > 0 (which was racy), we instead check to see if the cgroup.memory_control file exists. We can be reasonably certain if this file doesn't exist, and this is the first event we've received on the fd,the cgroup was cleaned up instead of an oom happening. This also matches what containerd does.

Signed-off-by: Peter Hunt <pehunt@redhat.com>