oom: check for leftover events before checking for deleted files
When we check for an oom, there are three cases we must worry about:
1. the cgroup OOM'd
2. the cgroup was destroyed
3. the cgroup OOM'd and was destroyed

We have always handled case 1 (when we get an oom, the eventfd gets an events, we write the oom file)
the second case is now being handled by the previous commit (we should not write oom files if the cgroup was only torn down)
the third case is handled in this commit. If both an oom and a destroy event happen, we need to make sure we properly set an oom. We can't use the handling code for the second case here, because the files will be gone, but an oom will have happened. Thus, before we check if the files exist, we need to check if there are still events in the eventfd. If there are, that means both oom and destroy happened, and we should write the oom files

Signed-off-by: Peter Hunt <pehunt@redhat.com>