ignore sigpipe
conmon getting killed from SIGPIPE is another way it can be spuriously killed. These situations are very hard to recover from, so we should ignore them.

This also fixes an issue where occasionally the splice() call to masterfd_stdin causes conmon to crash. It's pretty racy to try to check if stdin is still open, so we should just not crash instead

Signed-off-by: Peter Hunt <pehunt@redhat.com>