conmon: avoid hanging on stdin
now that we dropped splice(2), we need to make sure the read/write
loop doesn't hang indefinitely conmon.

Attempting to write to a fd that is not ready for write causes conmon
to block on the write(2) call, at the same time the container process
might be blocked attempting to write to stderr but since conmon is not
processing any data from the container, both processes are dead
locked.

Keep in a buffer what we read from the pipe and write to the
masterfd_stdin only when it is ready for write.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>