conn_sock: close stdout side when we don't close stdin
the main consumer of leave-stdin-open, cri-o, needs stdin to stay open, but also wants to consume all data from stdout.
It does not know when the output from stdout has ended unless conmon closes its side of the pipe.

the problem is, if we don't close masterfd_stdin, we also don't close masterfd_stdout, as they point to the same side pipe, which results in cri-o sitting and waiting for input forever.

instead, we need to close the stdout side of the pipe if we don't close masterfd_stdin

Signed-off-by: Peter Hunt <pehunt@redhat.com>