Update log-path parsing and unit-tests to handle a lot more edge cases.
When using strtok, you cannot tell if the original string had a : or not, because with

with "journald" you get driver = "journald" and path = NULL
with "journald:" you get driver = "journald" and path = NULL

That means invalid: writes to a file invalid, instead of throwing an error.
Similarly, when passing none or none: we make sure no file is written to disk.
Add debugging output for logdriver diagnostics and use in unit tests.
And "" and : just segfault.

All fixed now.

Signed-off-by: Joseph Gooch <mrwizard@dok.org>