attach: always chdir
the age of symlinking the attach socket to generate the sunpath must end.
The path is simply too long, and it is quite hacky anyway.

when setting up the attach socket, chdir to the directory before linking the sunpath.
That allows us to drop this age old hack, and future proof the code

Signed-off-by: Peter Hunt <pehunt@redhat.com>