conmon: store open FDs and close only them
now that we use a delay to call the cleanup program, we might end up
in a race where the event fd used by glib is close'd and it causes the
glib event handler to keep polling the closed file descriptor in a
tight loop.

To avoid closing files that are handled by glib, store what FDs are
opened when conmon first started and close only them.

Closes: https://github.com/containers/conmon/issues/221

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>