conn_sock: add full-attach option
unfortunately, older podman versions rely on conmon truncating the unix socket length when setting up the attach socket.

Introduce a flag to gate the changed behavior, allowing newer podman versions to be able to handle longer UID in rootless mode,
but also allowing older podman versions to work.

This also removes the logic for `do_chdir` in favor of using a /proc/self/fd entry for the sock path.
That allows conmon to not chdir before attaching to the unix socket.

Signed-off-by: Peter Hunt <pehunt@redhat.com>