Match `read(2)` buffer size to pipe size
Address a number of conditions flagged in issue #262.

The general thrust of this commit is to move closer to a one-to-one,
`read(2)` system call to `writev(2)` system call ratio by creating a
memory buffer for reading sized to match the stdio pipes.

The commit also allocates a maximum number of I/O vectors sized to the
read buffer size, calculating 2 I/O vectors per line, defaulting to a
target line size of 25 bytes (defined as the constant,
`AVG_LOG_LINE_TARGET`).

The commit also now allocates the I/O vectors and read buffer memory
using `mmap()` so that we don't cross page boundaries for I/O
operations.

As a result, the commit removes the `STDIO_BUF_SIZE` constance, using
`fcntl(F_GETPIPE_SZ)` calls to size all buffers for reading from pipes.

The `stdpipe_t` value is now passed to `write_to_remote_consoles()` so
that we do not have to play games with pre-/post- byte allocations to
buffers.