exit: fix race zombie reaper
fix a race condition where all the processes spawned by the OCI
runtime are not terminated yet and reap_children() would not clean
them up.

If an --exit-delay is specified, these proceses would be left
uncleaned until the delay expires.

When used for an exec session, the zombie process would prevent the
container namespaces to be cleaned up.

Closes: https://github.com/containers/conmon/issues/290

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>