conmon: free userdata files before exec cleanup
make sure all files under the userdata directory are closed before
calling the cleanup program.

This is needed when running on NFS since unlinking a file on NFS and
keeping it open causes a "silly rename" to keep a reference to the
file and doesn't permit rm -rf to cleanup the directory correctly.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>