Cirrus: Add meta task to keep CI VM images alive
This task is absolutely critical, as despite it being replaced, conmon
is still in use by many.  Having a meta task running on important
branches is needed to maintain a functioning CI.  That said, I also
manually marked the `c6431352024203264` series images for permanent
retention as well (for the `main` branch).

Signed-off-by: Chris Evich <cevich@redhat.com>