Fix tools/Makefile with GNU make 4.4
Before GNU make 4.4, "export" variables were not actually exported to $(shell)
commands, so the "export GO111MODULE=off" line had no effect. With
make >= 4.4, this variable is now set in the environment of the "go build"
command, which breaks it. Just remove the wrong variable assignment to have
the same behavour as previously.

Signed-off-by: Fabian Vogt <fvogt@suse.de>