logging: avoid opening /dev/null for each write
This patch improves the efficiency of the journald logging by removing
the usage of /dev/null in the write_journald function.

Add a new function which directly appends the segment without flushing
the buffer, so there is no need to pass a fd argument.

Another benefit is that now we get an error message instead of
silently ignoring what was previously stored in the buffer.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>