ensure console socket buffers are properly sized
On FreeBSD, the default socket buffer size for unix domain sockets is
just 8192 which is too small for largest message used on the console
socket which is 8193.  This causes output for processes which emit large
amounts of data to be lost on FreeBSD (for an example, see podman's
'podman exec/run - missing output' system test).

To avoid the problem, this commit increases the buffer sizes to
CONN_SOCK_BUF_SIZE (32768) which allows for several max-sized messages
to be in-flight.

Signed-off-by: Doug Rabson <dfr@rabson.org>