logging: Respect log-size-max immediately after open
When a container keeps restarting, conmon restarts and opens the log
file with O_APPEND every single time, resetting bytes_written to 0
with a non-empty file.
If the container fails before writting log-size-max bytes then the
log keeps growing indefinitely, potentially using up all the available
space.

This should be kept in check: initialize bytes_written to the file's
current size, so that it can be reset if required (it's possible that
the file gets re-open before the first write if it was already above the
limit, but that should almost never happen in practice)

In order to make this simpler make bytes written global variables

Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>