Make 'docs' target not depend on 'install.tools' if GOMD2MAN is set
If the GOMD2MAN (environment) variable is set, then we do not need the
'docs' target to depend on 'install.tools', which will build
gomd2man.

This avoids gomd2man to be build, even though the user already request
the usage of a particular gomd2man binary. It also removes an implicit
build dependency on go. The latter causes downstream bugs like
https://bugs.gentoo.org/931645

Signed-off-by: Florian Schmaus <flo@geekplace.eu>