include the correct header for syscall definitions
This file needs to utilize `__NR_seccomp`, which is defined in the linux
uapi headers, not seccomp.h, even though seccomp.h does itself
indirectly cause this header to be included as well.

Nothing else in this particular file needs seccomp.h so drop that
include altogether since it's now entirely unused.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>