Remove old vendored go-md2man
The version of go-md2man is quite old, and is not really need to be in
this repository. It is only used by `make docs` which does not make
sense to run in CI, as we don't check the resulting man page.

PS if needed, it is easy to install by this one-liner:

	go install github.com/cpuguy83/go-md2man/v2@latest

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>