Remove hack/tree_status.sh
The sole reason of having this was to produce a nice informative error
message in case of failure (and because the message was not explicitly
specified, it was entirely wrong in 1 of 2 use cases).

Let's drop it and just use git to fail the job if a tree is dirty.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>