Show the code coverage in `make test-coverage` output.
This commit adds the `make test-coverage` to recompile the conmon
with `--coverage` flag so it generates the `.gcno and `.gcda` (these
files are included in the `.gitignore` now) while running the tests.

After that, the `gcovr` is executed to show the code coverage based on
those files.

If `gcovr` is not found, it fallbacks to `gcov`.

Signed-off-by: Jan Kaluza <jkaluza@redhat.com>