Really run the container in the 04-runtime.bats.
Previously, the conmon has been tested in the 04-runtime.bats, but
we never tried to really run the container defined by the conmon.
It was therefore not clear whether the conmon configure the runc
correctly and whether the logs are really forwarded where they should
be.

This commit changes it by running the `runc start`. The `sleep` based
waiting login is replaced by polling of `runc state` to detect the
container `created` and `stopped` states - this makes tests faster
and more reliable.

The `assert` helper method to compare strings is borrowed from
the podman project as it is.

The `container cleanup on completion` test is removed, since it does
not test anything not tested by the previous tests.

Signed-off-by: Jan Kaluza <jkaluza@redhat.com>