Use the actual pipe size instead of 8,192
The default for linux pipes is typically something much larger than 8,192 bytes.
We take advantage of this to fact to avoid multiple rounds of reading from a
full pipe.

The constant, `STDIO_BUF_SIZE` is removed in favor of `fcntl(F_GETPIPE_SZ)`
calls for pipes / fifos, an adding `DEF_STDIO_BUF_SIZE` for the case where
`stdout` is not a PIPE (perhaps character special file) using a 64K buffer to
consume more data per system when possible.

Signed-off-by: Peter Portante <peter.portante@redhat.com>