Restore use of `writev()` system call
The use of the sequential `write()` system calls to replace `writev()` in commit
29d17be might be simpler to understand, but can lead to unwanted CPU overhead
when a high frequency of newlines exists in the data.  The original bugfix for
the log file corruption was the proper reset of the `iovcnt` on error.  We
restore the reset but always do it as part of the `writev_buffer_flush()` method
which avoids "leaking" knowledge of how it works.

Signed-off-by: Peter Portante <peter.portante@redhat.com>