Remove need for extra bytes for remote consoles
We use `writev()` for remote consoles to remove the need for extra bytes
surrounding the data to be written.

We really don't want to "leak" the protocol requirement for remote consoles to
the rest of the code base.  To hide that, and avoid two `write_all()` method
calls in a row, we promote `writev_buffer_flush()` to a `utils` method called
`writev_all()` to leverage I/O vectors in one call (memory copies will need to
occur for remote sockets, so using `writev()` avoids that).

Further, we use the `g_ptr_array_foreach()` method instead of doing it
ourselves to avoid breaking the `GPtrArray` encapsulation.

Signed-off-by: Peter Portante <peter.portante@redhat.com>