Remove need for extra bytes for remote consoles
Remote consoles are not always attached to a `conmon` process.  Don't bother
adding the "pre-byte" for the pipe, just add a `write()` system call for that
byte.  We also remove the null termination since all the operations must be done
by count (null bytes can be read from the pipe).

Signed-off-by: Peter Portante <peter.portante@redhat.com>