Use `size_t` with name `buflen` consistently
A few of the parameters were named `num_read` "leaking" the use it was named
for in a calling function.  We use `buflen` with `buf` more consistently, and
use `size_t` were possible since we don't need or want to pass a negative value
for the size of a buffer.

Signed-off-by: Peter Portante <peter.portante@redhat.com>